+++ Changelog.md
@@ -1,5 +1,9 @@
 ## 0.1.0
-*INCOMPLETE*
++ Added `beta.DFlashSpeculativeDecoding` type for configuring DFlash speculative decoding for LLM inference.
++ Added `beta.Eagle3SpeculativeDecoding` type for configuring DFlash speculative decoding for LLM inference.
++ Added `beta.SGLangInferenceMetadata.speculative_decoding` parameter for specifying speculative decoding models for accelerated LLM inference.
++ Added `beta.SGLangInferenceMetadata.max_running_requests` parameter for configuring SGLang engine concurrency.
++ Updated `BatchConfig.capacity` parameter to be optional for `continuous` batching.
 
 ## 0.0.104
 + Added `beta.CompileResource.from_hf_hub` method for defining compile resources from Huggingface Hub repositories.